bug: refactored product state
