+++ client/store/products.js
@@ -2,21 +2,17 @@ import axios from 'axios'
 
 //Action Types
 const ALL_PRODUCTS = 'ALL_PRODUCTS'
-const SINGLE_PRODUCT = 'SINGLE_PRODUCT'
 const DELETE_PRODUCT = 'DELETE_PRODUCT'
 const SALE_PRODUCTS = 'SALE_PRODUCTS'
 const TAG_PRODUCTS = 'TAG_PRODUCTS'
 const ADD_PRODUCT = 'ADD_PRODUCT'
-const UPDATE_PRODUCT = 'UPDATE_PRODUCT'
 
 //Action Creators
 const viewProducts = products => ({type: ALL_PRODUCTS, products})
-const singleProduct = product => ({type: SINGLE_PRODUCT, product})
 const deletingProduct = productId => ({type: DELETE_PRODUCT, productId})
 const saleProducts = products => ({type: SALE_PRODUCTS, products})
 const tagProducts = product => ({type: TAG_PRODUCTS, product})
 const addNewProduct = product => ({type: ADD_PRODUCT, product})
-const updatingProduct = products => ({type: UPDATE_PRODUCT, products})
 
 //Thunk Creator
 export const getAllProducts = () => {
@@ -24,17 +20,6 @@ export const getAllProducts = () => {
     try {
       const result = await axios.get('/api/products')
       dispatch(viewProducts(result.data))
-    } catch (err) {
-      console.error(err)
-    }
-  }
-}
-
-export const getSingleProduct = product => {
-  return async dispatch => {
-    try {
-      const result = await axios.get(`/api/products/${product}`)
-      dispatch(singleProduct(result.data))
     } catch (err) {
       console.error(err)
     }
@@ -85,24 +70,10 @@ export const addProduct = product => {
   }
 }
 
-export const updateProduct = (update, product) => {
-  return async dispatch => {
-    try {
-      console.log(update)
-      const result = await axios.put(`/api/products/${product}`, update)
-      dispatch(updatingProduct(result.data))
-    } catch (err) {
-      console.error(err)
-    }
-  }
-}
-
 //Reducer
 const manageProducts = (state = [], action) => {
   switch (action.type) {
     case ALL_PRODUCTS:
-      return action
-    case SINGLE_PRODUCT:
       return action
     case ADD_PRODUCT:
       state.products.push(action.product)
@@ -115,8 +86,6 @@ const manageProducts = (state = [], action) => {
         })
       }
     case SALE_PRODUCTS:
-      return action
-    case UPDATE_PRODUCT:
       return action
     case TAG_PRODUCTS:
       return action
